fix some warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3787 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/fronthlp.c
@@ -221,8 +221,8 @@ void identify_rom(const char* name, const char* hash, int length)
 	int found = 0;
 
 	/* remove directory name */
-	size_t i;
-	for (i = strlen(name)-1;i >= 0;i--)
+	int i;
+	for (i = (int)strlen(name)-1;i >= 0;i--)
 	{
 		if (name[i] == '/' || name[i] == '\\')
 		{
